Move queries upstream, to main NavBar component
