+++ app/core/components/NavbarDropdown.tsx
@@ -1,29 +1,15 @@
 import { Fragment, useState } from "react"
-import { Listbox, Transition, Dialog } from "@headlessui/react"
-import { Link, Routes, useMutation, useSession, useQuery, useRouter } from "blitz"
-import { Suspense } from "react"
-import { CheckIcon, MenuIcon, PlusSmIcon, SelectorIcon } from "@heroicons/react/solid"
-import { Close32, Menu32 } from "@carbon/icons-react"
-import { useCurrentUser } from "../hooks/useCurrentUser"
-import { useCurrentWorkspace } from "../hooks/useCurrentWorkspace"
+import { Transition, Dialog } from "@headlessui/react"
+import { Link, Routes, useMutation } from "blitz"
+import { Add32, Close32, Menu32 } from "@carbon/icons-react"
 import logout from "../../auth/mutations/logout"
 import SettingsModal from "../modals/settings"
-import changeSessionWorkspace from "../../workspaces/mutations/changeSessionWorkspace"
-import getDrafts from "../queries/getDrafts"
 import ResearchEqualsLogo from "./ResearchEqualsLogo"
-import { BellIcon } from "@heroicons/react/outline"
 import QuickDraft from "../../modules/components/QuickDraft"
-import getInvitedModules from "app/workspaces/queries/getInvitedModules"
 import DropdownNotificationModal from "../modals/DropdownNotificationModal"
 
-const DropdownContents = () => {
-  const currentUser = useCurrentUser()
-  const currentWorkspace = useCurrentWorkspace()
+const DropdownContents = ({ currentUser, currentWorkspace, router, invitedModules, drafts }) => {
   const [logoutMutation] = useMutation(logout)
-  const session = useSession()
-  const router = useRouter()
-  const [invitedModules] = useQuery(getInvitedModules, { session })
-  const [drafts] = useQuery(getDrafts, { session })
 
   if (currentUser && currentWorkspace) {
     return (
@@ -32,7 +18,7 @@ const DropdownContents = () => {
           <QuickDraft
             buttonText={
               <>
-                <PlusSmIcon
+                <Add32
                   className="inline w-4 h-4 fill-current text-indigo-500 dark:text-gray-400"
                   aria-hidden="true"
                 />
@@ -128,7 +114,7 @@ const DropdownContents = () => {
   }
 }
 
-const NavbarDropdown = () => {
+const NavbarDropdown = ({ currentUser, currentWorkspace, router, invitedModules, drafts }) => {
   let [isOpen, setIsOpen] = useState(false)
 
   return (
@@ -152,7 +138,7 @@ const NavbarDropdown = () => {
             }}
           >
             <span className="sr-only">Open menu</span>
-            <MenuIcon className="block h-6 w-6 text-gray-400 dark:text-gray-200 hover:bg-gray-100 hover:text-gray-500 dark:hover:bg-gray-800 rounded-md focus:ring-2 focus:ring-offset-0 focus:ring-gray-200" />
+            <Menu32 className="block h-6 w-6 text-gray-400 dark:text-gray-200 hover:bg-gray-100 hover:text-gray-500 dark:hover:bg-gray-800 rounded-md focus:ring-2 focus:ring-offset-0 focus:ring-gray-200" />
           </button>
         </>
       )}
@@ -209,9 +195,13 @@ const NavbarDropdown = () => {
                   </div>
                 </Dialog.Title>
                 <div className="items-center justify-between">
-                  <Suspense fallback="Loading...">
-                    <DropdownContents />
-                  </Suspense>
+                  <DropdownContents
+                    currentUser={currentUser}
+                    currentWorkspace={currentWorkspace}
+                    router={router}
+                    invitedModules={invitedModules}
+                    drafts={drafts}
+                  />
                 </div>
               </div>
             </Transition.Child>
